Rename forgeops install to forgeops apply
1. Rename `forgeops install` to `forgeops apply`
1. Add kubernetes flags only in the required cmds

Closes #16
