+++ docs/forgeops.md
@@ -16,9 +16,9 @@ forgeops is a tool for managing ForgeRock platform deployments
 
 ### SEE ALSO
 
+* [forgeops apply](forgeops_apply.md)	 - Install common platform components
 * [forgeops docs](forgeops_docs.md)	 - Generates docs
 * [forgeops doctor](forgeops_doctor.md)	 - Diagnose common cluster and platform deployments
-* [forgeops install](forgeops_install.md)	 - Install common platform components
 * [forgeops version](forgeops_version.md)	 - Print the version number of ForgeOps CLI tool
 
-###### Auto generated by spf13/cobra on 9-Dec-2020
+###### Auto generated by spf13/cobra on 10-Dec-2020
