+++ docs/forgeops_apply.md
@@ -1,26 +1,14 @@
-## forgeops install sa
+## forgeops apply
 
-Installs the ForgeRock secret-agent
+Install common platform components
 
 ### Synopsis
 
 
-	Installs the ForgeRock secret-agent:
-	  * Applies the latest secret-agent manifest
-	  * use --tag to specify a specific secret-agent version to install
-	
-
-```
-forgeops install sa [flags]
-```
+	Apply common platform components
+    
 
 ### Options
-
-```
-  -h, --help   help for sa
-```
-
-### Options inherited from parent commands
 
 ```
       --as string                      Username to impersonate for the operation
@@ -31,13 +19,14 @@ forgeops install sa [flags]
       --client-key string              Path to a client key file for TLS
       --cluster string                 The name of the kubeconfig cluster to use
       --context string                 The name of the kubeconfig context to use
+  -h, --help                           help for apply
       --insecure-skip-tls-verify       If true, the server's certificate will not be checked for validity. This will make your HTTPS connections insecure
       --kubeconfig string              Path to the kubeconfig file to use for CLI requests.
   -n, --namespace string               If present, the namespace scope for this CLI request
       --password string                Password for basic authentication to the API server
       --request-timeout string         The length of time to wait before giving up on a single server request. Non-zero values should contain a corresponding time unit (e.g. 1s, 2m, 3h). A value of zero means don't timeout requests. (default "0")
   -s, --server string                  The address and port of the Kubernetes API server
-  -t, --tag string                     Tag/version to install
+  -t, --tag string                     Tag/version to apply
       --tls-server-name string         Server name to use for server certificate validation. If it is not provided, the hostname used to contact the server is used
       --token string                   Bearer token for authentication to the API server
       --user string                    The name of the kubeconfig user to use
@@ -46,6 +35,8 @@ forgeops install sa [flags]
 
 ### SEE ALSO
 
-* [forgeops install](forgeops_install.md)	 - Install common platform components
+* [forgeops](forgeops.md)	 - forgeops is a tool for managing ForgeRock platform deployments
+* [forgeops apply quickstart](forgeops_apply_quickstart.md)	 - Installs the ForgeRock Cloud Deployment Quickstart (CDQ)
+* [forgeops apply sa](forgeops_apply_sa.md)	 - Installs the ForgeRock secret-agent
 
-###### Auto generated by spf13/cobra on 9-Dec-2020
+###### Auto generated by spf13/cobra on 10-Dec-2020
