+++ docs/forgeops_apply_quickstart.md
@@ -1,14 +1,26 @@
-## forgeops install
+## forgeops apply quickstart
 
-Install common platform components
+Installs the ForgeRock Cloud Deployment Quickstart (CDQ)
 
 ### Synopsis
 
 
-	Install common platform components
-    
+	Installs the ForgeRock Cloud Deployment Quickstart (CDQ):
+	  * Applies the latest quickstart manifest
+	  * use --tag to specify a specific CDQ version to install
+	
+
+```
+forgeops apply quickstart [flags]
+```
 
 ### Options
+
+```
+  -h, --help   help for quickstart
+```
+
+### Options inherited from parent commands
 
 ```
       --as string                      Username to impersonate for the operation
@@ -19,14 +31,13 @@ Install common platform components
       --client-key string              Path to a client key file for TLS
       --cluster string                 The name of the kubeconfig cluster to use
       --context string                 The name of the kubeconfig context to use
-  -h, --help                           help for install
       --insecure-skip-tls-verify       If true, the server's certificate will not be checked for validity. This will make your HTTPS connections insecure
       --kubeconfig string              Path to the kubeconfig file to use for CLI requests.
   -n, --namespace string               If present, the namespace scope for this CLI request
       --password string                Password for basic authentication to the API server
       --request-timeout string         The length of time to wait before giving up on a single server request. Non-zero values should contain a corresponding time unit (e.g. 1s, 2m, 3h). A value of zero means don't timeout requests. (default "0")
   -s, --server string                  The address and port of the Kubernetes API server
-  -t, --tag string                     Tag/version to install
+  -t, --tag string                     Tag/version to apply
       --tls-server-name string         Server name to use for server certificate validation. If it is not provided, the hostname used to contact the server is used
       --token string                   Bearer token for authentication to the API server
       --user string                    The name of the kubeconfig user to use
@@ -35,8 +46,6 @@ Install common platform components
 
 ### SEE ALSO
 
-* [forgeops](forgeops.md)	 - forgeops is a tool for managing ForgeRock platform deployments
-* [forgeops install quickstart](forgeops_install_quickstart.md)	 - Installs the ForgeRock Cloud Deployment Quickstart (CDQ)
-* [forgeops install sa](forgeops_install_sa.md)	 - Installs the ForgeRock secret-agent
+* [forgeops apply](forgeops_apply.md)	 - Install common platform components
 
-###### Auto generated by spf13/cobra on 9-Dec-2020
+###### Auto generated by spf13/cobra on 10-Dec-2020
