+++ docs/forgeops_apply_sa.md
@@ -1,23 +1,23 @@
-## forgeops install quickstart
+## forgeops apply sa
 
-Installs the ForgeRock Cloud Deployment Quickstart (CDQ)
+Installs the ForgeRock secret-agent
 
 ### Synopsis
 
 
-	Installs the ForgeRock Cloud Deployment Quickstart (CDQ):
-	  * Applies the latest quickstart manifest
-	  * use --tag to specify a specific CDQ version to install
+	Installs the ForgeRock secret-agent:
+	  * Applies the latest secret-agent manifest
+	  * use --tag to specify a specific secret-agent version to install
 	
 
 ```
-forgeops install quickstart [flags]
+forgeops apply sa [flags]
 ```
 
 ### Options
 
 ```
-  -h, --help   help for quickstart
+  -h, --help   help for sa
 ```
 
 ### Options inherited from parent commands
@@ -37,7 +37,7 @@ forgeops install quickstart [flags]
       --password string                Password for basic authentication to the API server
       --request-timeout string         The length of time to wait before giving up on a single server request. Non-zero values should contain a corresponding time unit (e.g. 1s, 2m, 3h). A value of zero means don't timeout requests. (default "0")
   -s, --server string                  The address and port of the Kubernetes API server
-  -t, --tag string                     Tag/version to install
+  -t, --tag string                     Tag/version to apply
       --tls-server-name string         Server name to use for server certificate validation. If it is not provided, the hostname used to contact the server is used
       --token string                   Bearer token for authentication to the API server
       --user string                    The name of the kubeconfig user to use
@@ -46,6 +46,6 @@ forgeops install quickstart [flags]
 
 ### SEE ALSO
 
-* [forgeops install](forgeops_install.md)	 - Install common platform components
+* [forgeops apply](forgeops_apply.md)	 - Install common platform components
 
-###### Auto generated by spf13/cobra on 9-Dec-2020
+###### Auto generated by spf13/cobra on 10-Dec-2020
